Finished the sim. Turns out there were some "pointer arithmetic" bugs, but now it passes
Note to self: my solution was kind of a kludge. I fiddled with wires in packetmem.v to
divide the forwader's address by 2, but I divided the packet length by 2 inside my
test forwarder. What's more is that the length is in words, not bytes, so I might need
to do some kind of fix for the CPU as well... and all this just makes me think that I
don't have a way to collect packets that are a weird number of bits (like 731 or
something).

Oh... what am I getting into?