In the middle of testing my new AXI infrastructure I moved the reg+strobe logic to a new module and made a testbench for it. It looks good. I haven't done a timing simulation yet.
I'm still in the middle of figureing out the easiest way to test the AXI register file
stuff. I may try instantiating an AXI VIP in a testbench (that way I can include the
testing code in the git repo). Aside from doing it the block diagram way, the other
options would be to manually twiddle the AXI wires (yuck) or write an HLS core that
drives the inputs (also yuck, but may be easier than the manual way).