A ton of minor bug fixes First, the super messy packet length logic gave me some trouble. I knew I should have just done it properly to begin with. Anyway, it's a little more consistent now.
There was also a typo in my test data for the sim. When I was expanding
it out to 64 bits, I forgot to set TLAST somewhere and the accepted
packet included a bunch of deadbeef stuff.

There may still be a few bugs to iron out, so I'm gonna take another look